Basic Modules for CRUD

+++ Owlylion.Klaws.Services/KlawsServiceBase.cs
@@ -1,17 +1,19 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.JsonPatch;
 using Owlylion.Klaws.Core;
+using Owlylion.Klaws.DataAccessLayer.Exceptions;
 using Owlylion.Klaws.DataAccessLayer.Repositories;
+using Owlylion.Klaws.Services.Interfaces;
+using Owlylion.Klaws.Services.Models;
 
 namespace Owlylion.Klaws.Services
 {
-    public class KlawsServiceBase<TDto, TModel> : IKlawsService<TDto>
-        where TDto : class, new()
-        where TModel : class, new()
+    public class KlawsServiceBase<TDataTransferObject, TModel> : IKlawsService<TDataTransferObject>
+        where TDataTransferObject : class, IKeyedObject, new()
+        where TModel : class, IKeyedObject, new()
     {
         protected readonly IKlawsRepository<TModel> Repository;
         protected readonly IMapper Mapper;
@@ -22,26 +24,64 @@ namespace Owlylion.Klaws.Services
             Mapper = mapper;
         }
 
-        public async Task Add(TDto dto)
+        public async Task<TDataTransferObject> Get(object key)
+        {
+            TModel model = await Repository.Get(key);
+            return Mapper.Map<TDataTransferObject>(model);
+        }
+
+        public async Task<PaginatedResult<TDataTransferObject>> List(int page, int perPage)
+        {
+            ICollection<TModel> modelList = await Repository.List(page, perPage);
+            ICollection<TDataTransferObject> dtoList = modelList
+                .Select(m => Mapper.Map<TDataTransferObject>(m))
+                .ToArray();
+            
+            return new PaginatedResult<TDataTransferObject>()
+            {
+                Page = page,
+                PerPage = perPage,
+                List = dtoList
+            };
+        }
+
+        public async Task<TDataTransferObject> Create(TDataTransferObject dto)
         {
             var model = Mapper.Map<TModel>(dto);
             await Repository.Add(model);
+            return Mapper.Map<TDataTransferObject>(model);
         }
 
-        public async Task<TDto> Get(object key)
+        public async Task<TDataTransferObject> Put(TDataTransferObject dto)
+        {
+            var model = Mapper.Map<TModel>(dto);
+            await Repository.Update(model);
+            return Mapper.Map<TDataTransferObject>(model);
+        }
+
+        public async Task<TDataTransferObject> Patch(object key, JsonPatchDocument<TDataTransferObject> dtoPatch)
         {
             TModel model = await Repository.Get(key);
-            return Mapper.Map<TDto>(model);
+            if (model == null)
+                throw new EntityNotFoundException($"Entity with Key {key} not found");
+
+            var dto = Mapper.Map<TDataTransferObject>(model);
+
+            dtoPatch.ApplyTo(dto);
+            var updatedModel = Mapper.Map<TModel>(dto);
+
+            await Repository.Update(updatedModel);
+            return dto;
         }
 
-        public Task Update(object key, Patch<TDto> patch)
+        public async Task Delete(object key)
         {
-            throw new NotImplementedException();
-        }
+            TModel model = await Repository.Get(key);
+            if (model == null)
+                throw new EntityNotFoundException($"Entity with Key {key} not found");
 
-        public async Task Remove(object key)
-        {
             await Repository.Remove(key);
+            await Repository.Save();
         }
     }
 }
